Move around reporting code to error module
