+++ core/src/error.rs
@@ -2,8 +2,12 @@
 //!
 //! Define error types for different phases of the execution, together with functions to generate a
 //! [codespan](https://crates.io/crates/codespan-reporting) diagnostic from them.
+use crate::cache::Cache;
+
 pub use codespan::{FileId, Files};
 pub use codespan_reporting::diagnostic::{Diagnostic, Label, LabelStyle};
+
+use codespan_reporting::term::termcolor::{ColorChoice, StandardStream};
 use lalrpop_util::ErrorRecovery;
 use malachite::num::conversion::traits::ToSci;
 
@@ -2280,4 +2284,63 @@ impl IntoDiagnostics<FileId> for ReplError {
             }
         }
     }
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub struct ColorOpt(pub(crate) clap::ColorChoice);
+
+impl From<clap::ColorChoice> for ColorOpt {
+    fn from(color_choice: clap::ColorChoice) -> Self {
+        Self(color_choice)
+    }
+}
+
+impl From<ColorOpt> for ColorChoice {
+    fn from(c: ColorOpt) -> Self {
+        use std::io::{stdout, IsTerminal};
+
+        match c.0 {
+            clap::ColorChoice::Auto => {
+                if stdout().is_terminal() {
+                    ColorChoice::Auto
+                } else {
+                    ColorChoice::Never
+                }
+            }
+            clap::ColorChoice::Always => ColorChoice::Always,
+            clap::ColorChoice::Never => ColorChoice::Never,
+        }
+    }
+}
+
+/// Pretty-print an error on stderr.
+///
+/// # Arguments
+///
+/// - `cache` is the file cache used during the evaluation, which is required by the reporting
+/// infrastructure to print point at specific locations and print snippet when needed.
+pub fn report<E: IntoDiagnostics<FileId>>(cache: &mut Cache, error: E, color_opt: ColorOpt) {
+    let stdlib_ids = cache.get_all_stdlib_modules_file_id();
+    report_with(cache.files_mut(), stdlib_ids.as_ref(), error, color_opt)
+}
+
+/// Report an error on `stderr`, provided a file database and a list of stdlib file ids.
+fn report_with<E: IntoDiagnostics<FileId>>(
+    files: &mut Files<String>,
+    stdlib_ids: Option<&Vec<FileId>>,
+    error: E,
+    color_opt: ColorOpt,
+) {
+    let writer = StandardStream::stderr(color_opt.into());
+    let config = codespan_reporting::term::Config::default();
+    let diagnostics = error.into_diagnostics(files, stdlib_ids);
+
+    let result = diagnostics
+        .iter()
+        .try_for_each(|d| codespan_reporting::term::emit(&mut writer.lock(), &config, files, d));
+
+    match result {
+        Ok(()) => (),
+        Err(err) => panic!("error::report_with(): could not print an error on stderr: {err}"),
+    };
 }
